rename to test_server to be more clear what this is
